requirements.txt
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=0.5
python-dotenv>=1.0
requests>=2.28
